byonoy: rename backend.py → driver.py + docs fixes
- pylabrobot/byonoy/backend.py → driver.py (and backend_tests.py →
  driver_tests.py). Imports updated; behavior unchanged.

Docs:
- index.md: add led_bar.ipynb to toctree (was orphaned)
- index.md: add Absorbance 96 models table (mirror of Luminescence 96
  table); strip the same table from absorbance_96/hello-world.ipynb
- luminescence_96/hello-world.ipynb: drop \"a heartbeat thread is running\"
  (ping worker was removed); fix api.version_no → api (request_api_version
  now returns int)
- luminescence_96/led_bar.ipynb: update Source: ref to driver.py; strip
  stale execution outputs that named the old module path / pid hex
