+++ docs/user_guide/byonoy/index.md
@@ -5,7 +5,18 @@
 
 absorbance_96/hello-world
 luminescence_96/hello-world
+luminescence_96/led_bar
 ```
+
+## Absorbance 96 models
+
+| Model | PLR resource | Factory function |
+|---|---|---|
+| A96A full setup | `ByonoyAbsorbance96` + illumination unit | `byonoy_a96a` |
+| Detection unit only | `ByonoyAbsorbance96` | `byonoy_a96a_detection_unit` |
+| Illumination unit | `Resource` | `byonoy_a96a_illumination_unit` |
+| Parking base (no backend) | `ByonoyAbsorbanceBaseUnit` | `byonoy_a96a_parking_unit` |
+| SBS adapter | `ResourceHolder` | `byonoy_sbs_adapter` |
 
 ## Luminescence 96 models
